Review: simplify table interface; move to infrastructure/
